New entrypoint

+++ entrypoint.sh
@@ -1,18 +1,36 @@
-certbot_renew() {
-    if certbot certonly -n --agree-tos --keep --expand --webroot -w /mnt/acme-webroot -m $LE_EMAIL $LE_DOMAIN_ARGS; then
-        nginx -s reload
-    else
-        echo "Renewal failed!"
+trap shutdown TERM
+trap 'exit 1' QUIT
+
+shutdown() {
+    if [ -f /run/nginx/nginx.pid ]; then
+        NGINX_PID=$(cat /run/nginx/nginx.pid)
+        kill $NGINX_PID
+        wait $NGINX_PID
     fi
+    exit 0
 }
 
-try_load_ssl_config() {
+certbot_renew() {
+    certbot certonly -n --agree-tos --keep --expand --webroot -w /mnt/acme-webroot -m $LE_EMAIL $LE_DOMAIN_ARGS
+
+    if [ "$LE_RENEW_ALL" = "true" ]; then
+        certbot renew -n
+    fi
+
+    ensure_ssl_config
+    nginx -s reload
+}
+
+ensure_ssl_config() {
     if [ -f /etc/letsencrypt/live/$LE_MAIN_DOMAIN/fullchain.pem ]; then
-        echo "ssl_certificate     /etc/letsencrypt/live/$LE_MAIN_DOMAIN/fullchain.pem;" > /etc/nginx/http.d/_cert.conf
-        echo "ssl_certificate_key /etc/letsencrypt/live/$LE_MAIN_DOMAIN/privkey.pem;"  >> /etc/nginx/http.d/_cert.conf
-        cp /etc/nginx/http.d.ssl/* /etc/nginx/http.d
-        pidof nginx && nginx -s reload
-        echo "SSL config loaded"
+        if [ -f /etc/nginx/http.d/_cert.conf ]; then
+            echo "SSL config already active"
+        else
+            echo "ssl_certificate     /etc/letsencrypt/live/$LE_MAIN_DOMAIN/fullchain.pem;" > /etc/nginx/http.d/_cert.conf
+            echo "ssl_certificate_key /etc/letsencrypt/live/$LE_MAIN_DOMAIN/privkey.pem;"  >> /etc/nginx/http.d/_cert.conf
+            cp /etc/nginx/http.d.ssl/* /etc/nginx/http.d
+            echo "SSL config activated"
+        fi
     else
         echo "SSL config SKIPPED"
     fi
@@ -24,27 +42,29 @@ for i in $LE_DOMAINS; do
 done
 
 if [ "$NGINX_LOGS" == true ]; then
-    ln -sf /dev/stdout /var/log/nginx/access.log
-    ln -sf /dev/stderr /var/log/nginx/error.log
+    # https://serverfault.com/a/932888
+    ln -sf /proc/1/fd/1 /var/log/nginx/access.log
+    ln -sf /proc/1/fd/2 /var/log/nginx/error.log
 else
     ln -sf /dev/null /var/log/nginx/access.log
     ln -sf /dev/null /var/log/nginx/error.log
 fi
 
-mkdir -p /run/nginx
+rm -f /run/nginx/nginx.pid
+
+ensure_ssl_config
 
 if [ "$NGINX_TEST" == true ]; then
-    try_load_ssl_config
     nginx -t
     exit $?
 fi
 
-nginx -g 'daemon off;' &
+nginx
 certbot_renew
 
-try_load_ssl_config
+sh /monitor.sh &
 
 while true; do
-    sleep 43200
+    sleep 43200 & wait $!
     certbot_renew
 done
